FW : Spec func "adjust Gvar with constant" bugfix
Negative values work now.
Limit is Gvar one : 120.
Thanks Guilm :+1:

+++ OpenAVRc/functions.cpp
@@ -264,7 +264,7 @@ void evalFunctions()
 #if defined(GVARS)
         case FUNC_ADJUST_GVAR:
           if (CFN_GVAR_MODE(cfn) == FUNC_ADJUST_GVAR_CONSTANT) {
-            SET_GVAR(CFN_GVAR_INDEX(cfn), CFN_PARAM(cfn), mixerCurrentFlightMode);
+            SET_GVAR(CFN_GVAR_INDEX(cfn), (int8_t)CFN_PARAM(cfn), mixerCurrentFlightMode);
           } else if (CFN_GVAR_MODE(cfn) == FUNC_ADJUST_GVAR_GVAR) {
             SET_GVAR(CFN_GVAR_INDEX(cfn), GVAR_VALUE(CFN_PARAM(cfn), getGVarFlightPhase(mixerCurrentFlightMode, CFN_PARAM(cfn))), mixerCurrentFlightMode);
           } else if (CFN_GVAR_MODE(cfn) == FUNC_ADJUST_GVAR_INC) {
